chore: rename context file
