+++ apps/extension/src/ui/apps/dashboard/layout/Sidebar/MainSidebar.tsx
@@ -17,7 +17,7 @@ import {
   StarIcon,
   ZapIcon,
 } from "@talismn/icons"
-import { useBuyTokensModal } from "@ui/domains/Asset/Buy/BuyTokensModalContext"
+import { useBuyTokensModal } from "@ui/domains/Asset/Buy/useBuyTokensModal"
 import { useSelectedAccount } from "@ui/domains/Portfolio/useSelectedAccount"
 import { useAnalytics } from "@ui/hooks/useAnalytics"
 import { useIsFeatureEnabled } from "@ui/hooks/useFeatures"
